requirements.txt
torch~=2.6.0
requests~=2.32.3
streamlit~=1.42.0
pandas~=2.2.3
numpy~=2.0.2
pillow~=11.1.0
torchvision~=0.21.0
scikit-learn~=1.6.1
clip~=1.0
beautifulsoup4~=4.13.3